// src/types/popular-movies.ts
import z from "zod";

export const POPULAR_MOVIE_VALIDATOR = z.object({
  adult: z.boolean(),
  backdrop_path: z.string().nullable(),
  genre_ids: z.array(z.number()),
  id: z.number(),
  original_language: z.string(),
  original_title: z.string(),
  overview: z.string(),
  popularity: z.number(),
  poster_path: z.string().nullable(),
  release_date: z.string(),
  title: z.string(),
  video: z.boolean(),
  vote_average: z.number(),
  vote_count: z.number(),
});

export const TOPRATED_MOVIE_VALIDATOR = POPULAR_MOVIE_VALIDATOR;